pyproject.toml
[tool.poetry]
name = "pycommerce"
version = "0.1.0"
description = "Ecommerce Example Application"
authors = ["alefeans <alefeans2@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = { extras = ["standard"], version = "^0.24.0" }
alembic = "^1.10.2"
pydantic = { extras = ["dotenv"], version = "^1.10.7" }
sqlmodel = "^0.0.12"
httpx = "^0.25.0"
asyncpg = "^0.29.0"
greenlet = "^3.0.1"
bcrypt = "^4.0.1"
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
python-multipart = "^0.0.6"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
mypy = "^1.1.1"
pytest = "^7.2.2"
pytest-watch = "^4.2.0"
pytest-asyncio = "^0.21.0"
pytest-cov = "^4.0.0"
ruff = "^0.1.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = "tests"
asyncio_mode = "auto"
filterwarnings = "ignore::DeprecationWarning"

[tool.poetry.scripts]
app = "pycommerce:start_server"

[tool.black]
line-length = 95

[tool.ruff]
line-length = 95
